+++ packages/terra-framework-docs/src/terra-dev-site/test/date-time-picker/DateTimePickerMinDate.test.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import DateTimePickerExampleTemplate from './common/DateTimePickerExampleTemplate';
+
+const DateTimePickerDefault = () => (
+  <DateTimePickerExampleTemplate value="1900-01-01T00:00" />
+);
+
+export default DateTimePickerDefault;