+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+## 1.8.1 - (March 29, 2022)
+
+* Added
+  * Added docs for `terra-date-time-picker`
+  * Added tests for `terra-date-time-picker`
+
 ## 1.8.0 - (March 15, 2022)
 
 * Changed